Add pids for a xtion revolute joint
A revolute joint with 0 limits needs to be defined for the Xtion pose wrt /base_link in order to get the point cloud properly generated in simulation. Using fixed joints is not supported by Gazebo and this is the typical workaround, i.e. defining a revolute joint with 0 limits and
fake PID parameters.